Moved the wasmboy options into the wasmboy file
