+++ demo/debugger/components/wasmboyInfo/wasmboyInfo.js
@@ -6,7 +6,7 @@ import { h, Component } from 'preact';
 import { Pubx } from 'pubx';
 import { PUBX_KEYS } from '../../pubx.config';
 
-import WasmBoy from '../../wasmboy';
+import { WasmBoy } from '../../wasmboy';
 
 import './wasmboyInfo.css';
