+++ demo/debugger/index.js
@@ -1,6 +1,6 @@
 import { h, render, Component } from 'preact';
 
-import WasmBoy from './wasmboy';
+import { WasmBoy } from './wasmboy';
 
 import loadScript from 'load-script';
