+++ demo/debugger/wasmboy.js
@@ -1,5 +1,71 @@
 // Single place to import WasmBoy
 // This is so we can swap between Different Lib outputs Easily
+import { Pubx } from 'pubx';
+import { PUBX_KEYS } from './pubx.config';
 
-import { WasmBoy } from '../../dist/wasmboy.wasm.esm';
-export default WasmBoy;
+export { WasmBoy } from '../../dist/wasmboy.wasm.esm';
+
+// Variables to tell if our callbacks were ever run
+let saveStateCallbackCalled = false;
+let graphicsCallbackCalled = false;
+let audioCallbackCalled = false;
+
+// WasmBoy Options
+export const WasmBoyDefaultOptions = {
+  isGbcEnabled: true,
+  isAudioEnabled: true,
+  frameSkip: 0,
+  audioBatchProcessing: true,
+  timersBatchProcessing: false,
+  audioAccumulateSamples: true,
+  graphicsBatchProcessing: false,
+  graphicsDisableScanlineRendering: false,
+  tileRendering: true,
+  tileCaching: true,
+  gameboyFrameRate: 60,
+  updateGraphicsCallback: imageDataArray => {
+    if (!graphicsCallbackCalled) {
+      console.log('Graphics Callback Called! Only Logging this once... imageDataArray:', imageDataArray);
+      graphicsCallbackCalled = true;
+    }
+  },
+  updateAudioCallback: (audioContext, audioBufferSourceNode) => {
+    if (!audioCallbackCalled) {
+      console.log(
+        'Audio Callback Called! Only Logging this once... audioContext, audioBufferSourceNode:',
+        audioContext,
+        audioBufferSourceNode
+      );
+      audioCallbackCalled = true;
+    }
+  },
+  saveStateCallback: saveStateObject => {
+    if (!saveStateCallbackCalled) {
+      console.log('Save State Callback Called! Only Logging this once... saveStateObject:', saveStateObject);
+      saveStateCallbackCalled = true;
+    }
+
+    // Function called everytime a savestate occurs
+    // Used by the WasmBoySystemControls to show screenshots on save states
+    const canvasElement = getCanvasElement();
+    if (canvasElement) {
+      saveStateObject.screenshotCanvasDataURL = getCanvasElement().toDataURL();
+    }
+  },
+  onReady: () => {
+    console.log('onReady Callback Called!');
+    Pubx.get(PUBX_KEYS.WASMBOY).update();
+  },
+  onPlay: () => {
+    console.log('onPlay Callback Called!');
+    Pubx.get(PUBX_KEYS.WASMBOY).update();
+  },
+  onPause: () => {
+    console.log('onPause Callback Called!');
+    Pubx.get(PUBX_KEYS.WASMBOY).update();
+  },
+  onLoadedAndStarted: () => {
+    console.log('onLoadedAndStarted Callback Called!');
+    Pubx.get(PUBX_KEYS.WASMBOY).update();
+  }
+};
